Extracts methods and rearranges code that is directly run
Some code which does not define a function is spread through out the
content.js file. This commit seeks to consolidate this code to a lower
number, like directly at the file head for constant declaration, or the
file end.

The main routine on top-level was extracted into the function
"runUncybered" which itself still showed a high number of encapsulation
of fors and ifs. That part that takes over the heavy lifting was
extracted into its own function for better overview.

One last inline function definition was replaced by a fat arrow function

Also some JSDoc comments have been placed. Further are to come.