Introduce test set-up with npm and mocha/chai as test runner suite
Previously the only way to check the browser extension was to install it
in Chrome. Now, automatized unit tests can now be run on the shell.

npm is used to handle javascript package dependencies which are solely
used to import the test runner framework mocha and the assertion frame-
work chai. As usual, an npm script can start the unit tests.

In order to declunk the productive code from the test code a new
directory structure main and test was created and the appropriate
content moved into their directory.

The global variable sourceRegexToTargetWords was removed from parameter
lists because it is globally available.

Add node_modules and diverse project files to .gitignore

README was converted into a markdown file.
Usage instructions are added to the new README.md.